Selections with CTRL+A will now work. Internal modifications. Constructor Rename
CTRL+A never used to work because we were selectively setting the selected state of jsh-char's to true only if they were false. This means that the animations that played for the caret blinking would be out of sync. Using a little trick we're able to force the browser to rerender the carets every time the update caret code is called.

Internal modifications include removing the possible branching in the updateCaret internal function.

Constructor for our object renamed to Jsh from Terminal.